// client/src/api/model/combatRollRequest.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Genesys API Documentation
 * Interactive API documentation
 * OpenAPI spec version: 1.0
 */

export interface CombatRollRequest {
  /** @minLength 1 */
  attackerId: string;
  /** @minLength 1 */
  weaponInstanceId: string;
  /** @minLength 1 */
  targetEnemyInstanceId: string;
  environmentalSetbackDice?: number;
  situationalBoostDice?: number;
}