// client/src/api/model/resolveChoicesRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Genesys API Documentation
 * Interactive API documentation
 * OpenAPI spec version: 1.0
 */
import type { SelectedChoice } from './selectedChoice';

export interface ResolveChoicesRequest {
  rollSessionId?: string;
  attackerId?: string;
  targetEnemyInstanceId?: string;
  selectedChoices?: SelectedChoice[];
}