// client/src/api/model/rollEvaluationResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Genesys API Documentation
 * Interactive API documentation
 * OpenAPI spec version: 1.0
 */
import type { GenesysSymbolResults } from './genesysSymbolResults';
import type { SpendOption } from './spendOption';

export interface RollEvaluationResponse {
  rollSessionId?: string;
  results?: GenesysSymbolResults;
  finalWoundsInflicted?: number;
  spendOptions?: SpendOption[];
}