// client/src/api/model/spendOption.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Genesys API Documentation
 * Interactive API documentation
 * OpenAPI spec version: 1.0
 */
import type { ResultType } from './resultType';

export interface SpendOption {
  name?: string;
  type?: ResultType;
  amount?: number;
  repeatable?: boolean;
}